renamed rChainService folder
