+++ server/modules/analysis/service/rChain/rCalculation.js
