+++ server/modules/analysis/service/rChain/rFunctions/arena.js
