+++ server/modules/analysis/service/rChain/rFunctions/utility.js
